+++ cpp/cpp05/ex03/Bureaucrat.cpp
@@ -6,7 +6,7 @@
 /*   By: bposa <bposa@student.hive.fi>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/04 22:09:06 by bposa             #+#    #+#             */
-/*   Updated: 2025/03/27 17:30:05 by bposa            ###   ########.fr       */
+/*   Updated: 2025/03/27 18:06:50 by bposa            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ Bureaucrat::~Bureaucrat() {}
 
 
 /*	Constructor Overload	*/
-Bureaucrat::Bureaucrat(const string& name, long long grade) : _name(name), _grade(grade) {
+Bureaucrat::Bureaucrat(const string& name, unsigned short grade) : _name(name), _grade(grade) {
 	if (grade < _bestGrade) {
 		throw GradeTooHighException();
 	}
